+++ src/ai/backend/cli/loader.py
@@ -23,12 +23,11 @@ def load_entry_points(
         else:
             prefix, _, subprefix = entrypoint.name.partition(".")
             try:
+                subcmd = entrypoint.load()
                 if not subprefix:
-                    subcmd = entrypoint.load()
                     main.add_command(subcmd, name=prefix)
                 else:
-                    subcmd = entrypoint.load()
                     main.commands[prefix].add_command(subcmd, name=subprefix)  # type: ignore
             except ImportError:
-                log.warning("Failed to import %r (%s)", entrypoint, prefix)
+                log.exception("Failed to import %r (%s)", entrypoint, prefix)
     return main
