+++ src/ai/backend/common/bgtask.py
@@ -26,6 +26,8 @@ from aiohttp_sse import sse_response
 from redis.asyncio import Redis
 from redis.asyncio.client import Pipeline
 
+from ai.backend.logging import BraceStyleAdapter
+
 from . import redis_helper
 from .events import (
     BgtaskCancelledEvent,
@@ -35,7 +37,6 @@ from .events import (
     EventDispatcher,
     EventProducer,
 )
-from .logging import BraceStyleAdapter
 from .types import AgentId, Sentinel
 
 sentinel: Final = Sentinel.TOKEN
