+++ src/ai/backend/common/distributed.py
@@ -6,7 +6,7 @@ from typing import TYPE_CHECKING, Callable, Final
 
 from aiomonitor.task import preserve_termination_log
 
-from .logging import BraceStyleAdapter
+from ai.backend.logging import BraceStyleAdapter
 
 if TYPE_CHECKING:
     from .events import AbstractEvent, EventProducer
