+++ src/ai/backend/common/docker.py
@@ -27,13 +27,14 @@ import trafaret as t
 import yarl
 from packaging import version
 
+from ai.backend.logging import BraceStyleAdapter
+
 from . import validators as tx
 from .arch import arch_name_aliases
 from .etcd import AsyncEtcd
 from .etcd import quote as etcd_quote
 from .etcd import unquote as etcd_unquote
 from .exception import InvalidImageName, InvalidImageTag, UnknownImageRegistry
-from .logging import BraceStyleAdapter
 from .service_ports import parse_service_ports
 
 __all__ = (
