+++ src/ai/backend/common/etcd.py
@@ -53,7 +53,8 @@ from etcd_client import (
     Txn as EtcdTransactionAction,
 )
 
-from .logging_utils import BraceStyleAdapter
+from ai.backend.logging import BraceStyleAdapter
+
 from .types import HostPortPair, QueueSentinel
 
 __all__ = (
