+++ src/ai/backend/common/etcd_etcetra.py
@@ -39,8 +39,9 @@ from etcetra.client import EtcdClient, EtcdTransactionAction
 from etcetra.types import CompareKey, EtcdCredential
 from etcetra.types import HostPortPair as EtcetraHostPortPair
 
+from ai.backend.logging import BraceStyleAdapter
+
 from .etcd import ConfigScopes
-from .logging_utils import BraceStyleAdapter
 from .types import HostPortPair, QueueSentinel
 
 __all__ = (
