+++ src/ai/backend/common/plugin/__init__.py
@@ -8,11 +8,11 @@ from typing import Any, ClassVar, Dict, Generic, Iterator, Mapping, Optional, Tu
 from weakref import WeakSet
 
 from ai.backend.common.asyncio import cancel_tasks
+from ai.backend.logging import BraceStyleAdapter
 from ai.backend.plugin.entrypoint import scan_entrypoints
 
 from ..etcd import AsyncEtcd
 from ..exception import ConfigurationError
-from ..logging import BraceStyleAdapter
 
 log = BraceStyleAdapter(logging.getLogger(__spec__.name))
