+++ src/ai/backend/common/plugin/hook.py
@@ -7,7 +7,8 @@ from typing import Any, Final, List, Optional, Protocol, Sequence, Tuple, Union
 
 import attrs
 
-from ..logging import BraceStyleAdapter
+from ai.backend.logging import BraceStyleAdapter
+
 from . import AbstractPlugin, BasePluginContext
 
 log = BraceStyleAdapter(logging.getLogger(__spec__.name))
