+++ src/ai/backend/manager/cli/agent.py
@@ -9,7 +9,7 @@ from alembic.config import Config
 
 from ai.backend.common.types import AgentId
 from ai.backend.logging import BraceStyleAdapter
-from ai.backend.logging_utils import enforce_debug_logging
+from ai.backend.logging.utils import enforce_debug_logging
 
 if TYPE_CHECKING:
     from .context import CLIContext
