+++ src/ai/backend/manager/cli/redis.py
@@ -10,7 +10,7 @@ import redis
 from ai.backend.common import redis_helper
 from ai.backend.common.types import RedisConnectionInfo
 from ai.backend.logging import BraceStyleAdapter
-from ai.backend.logging_utils import enforce_debug_logging
+from ai.backend.logging.utils import enforce_debug_logging
 from ai.backend.manager.cli.context import redis_ctx
 
 if TYPE_CHECKING:
